Heroku deployment version
Added files and changed configuation configuration files as necessary
for deployment of the app to Heroku. Deployed app to Heroku and
verified with unittests. Created script to initialize the database
for Postman tests. Finished Postman authorization tests and exported
with a version to run locally and a version to run against Heroku.
Verified authorization tests against Heroku deployment.

Finished Postman authorization
tests with a version for localhost and a version for heroku de

    TODO:
        - Create app documentation